move Place Search bar to a separate component to encapsulate state and better perf
